Updated TMC file with correct TMC Codes
The previous TMC codes were based on the second table from the link on
the top of tmc.py, which turned out to be incorrect since the code went
past 255.

I am unsure about some details about these codes, such as what
is the second table for, and what are the other columns in the first
table (assumed to be the TMC table) besides the code and description
column. Hopefully more details will be discovered in the future.